Added AssignmentController, TagController, ParticipationController

+++ Havoc-API/Havoc-API/Controllers/TagController.cs
@@ -3,6 +3,8 @@ using Havoc_API.Exceptions;
 using Havoc_API.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Havoc_API.Exceptions;
+using Havoc_API.Models;
 
 namespace Havoc_API.Controllers;
 
@@ -13,11 +15,15 @@ namespace Havoc_API.Controllers;
 [ApiController]
 public class TagController : ControllerBase
 {
+    private readonly IUserService _userService;
     private readonly ITagService _tagService;
+    private readonly IParticipationService _participationService;
 
-    public TagController(ITagService tagService)
+    public TagController(ITagService tagService, IUserService userService, IParticipationService participationService)
     {
         _tagService = tagService;
+        _userService = userService;
+        _participationService = participationService;
     }
 
     [HttpGet]
@@ -28,16 +34,30 @@ public class TagController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> AddTagToTaskAsync([FromBody] TagPOST tag, [FromRoute] int taskId)
+    public async Task<IActionResult> AddTagsToTaskAsync(IEnumerable<TagPOST> tags, int taskId, int projectId)
     {
-        var newTagId = await _tagService.AddTagToTaskAsync(tag, taskId);
-        return Created($"api/tasks/{taskId}/tags/{newTagId}", new { TagId = newTagId });
+
+        var role =
+        await _participationService
+            .GetUserRoleInProjectAsync(_userService.GetUserId(Request), projectId);
+
+        if (!role.CanEditTask())
+            return Unauthorized("You have no permission to edit task");
+
+        var result = await _tagService.AddTagsToTaskAsync(tags, taskId, projectId);
+        return Ok(result);
 
     }
 
-    [HttpDelete("{tagId}")]
-    public async Task<IActionResult> DeleteTagFromTaskAsync([FromRoute] int taskId, [FromRoute] int tagId)
+    [HttpDelete]
+    public async Task<IActionResult> DeleteTagFromTaskAsync( int taskId, int tagId, int projectId)
     {
+            var role =
+        await _participationService
+            .GetUserRoleInProjectAsync(_userService.GetUserId(Request), projectId);
+
+        if (!role.CanEditTask())
+            return Unauthorized("You have no permission to edit task");
         await _tagService.DeleteTagFromTaskAsync(tagId, taskId);
         return NoContent();
     }
